+++ tests/Feature/CollectionEngineTest.php
@@ -44,6 +44,13 @@ class CollectionEngineTest extends TestCase
         $this->assertCount(1, $models);
         $this->assertEquals(1, $models[0]->id);
 
+        $models = SearchableUserModel::search('Taylor')->query(function ($query) {
+            $query->where('email', 'like', 'taylor@laravel.com');
+        })->get();
+
+        $this->assertCount(1, $models);
+        $this->assertEquals(1, $models[0]->id);
+
         $models = SearchableUserModel::search('Abigail')->where('email', 'abigail@laravel.com')->get();
         $this->assertCount(1, $models);
         $this->assertEquals(2, $models[0]->id);
@@ -53,6 +60,9 @@ class CollectionEngineTest extends TestCase
 
         $models = SearchableUserModel::search('Taylor')->where('email', 'taylor@laravel.com')->get();
         $this->assertCount(1, $models);
+
+        $models = SearchableUserModel::search('otwell')->get();
+        $this->assertCount(2, $models);
 
         $models = SearchableUserModel::search('laravel')->get();
         $this->assertCount(2, $models);